+++ docs/components/Markdown.tsx
@@ -29,5 +29,5 @@ export default ({ content }: MarkdownProps) => {
     }
   }, content);
 
-  return <div class="markdown-container w-100%" ref={ref}></div>;
+  return <div class="markdown-container w-100%" ref={ref} />;
 };